+++ nc-functions.php
@@ -11,7 +11,6 @@
  * 	- Initialize
  *  - Role Capabilities
  * 	- Google Analytics
- *  - Custom list authors
  *
  * @package WordPress
  * @package NursingClioHelperFunctions
@@ -20,6 +19,7 @@
  * @copyright Copyright (c) 2015, Adam Turner, GPL 2.0+
  * @link https://github.com/admturner/nc-functions-wp-plugin
  * @license http://www.gnu.org/licenses/gpl.html
+
  * 
  * @todo Add function_exists() wrappers
  * @todo Make more robust (see http://solislab.com/blog/plugin-activation-checklist/)